requirements.txt
python-telegram-bot==21.5
openai==1.40.0
apscheduler==3.10.4
pytz==2024.1
